requirements.txt
Flask
pymongo[srv]
pytest
pytest-cov
PyJWT
Werkzeug
configparser
python-dotenv
